+++ CHANGES.md
@@ -5,6 +5,7 @@
 - **Breaking:** Update to GAP 4.16.0
 - Update the GAP package distribution to 4.16.0
 - Update the `utils` GAP package to 0.96
+- Improve GAP-originated exceptions to include structured GAP stack traces
 
 ## Version 0.16.7 (released 2026-06-09)